+++ tests/ProdManager/helpers/test_helpers_calendar.py
@@ -35,3 +35,28 @@ def test_calendar():
     ics_data = calendar.render()
 
     assert "BEGIN:VCALENDAR" in str(ics_data)
+    assert "SUMMARY:[CHG0000000] TEST maintenance" in str(ics_data)
+
+def test_calendar():
+  with app.app_context():
+    maintenance = Maintenance(
+      id = 1,
+      name = "TEST maintenance",
+      description = "The wonderfull description\nsecond line",
+      status = MaintenanceStatus.SCHEDULED,
+      service_status = ServiceStatus.DOWN,
+      creation_date = current_date(),
+      scheduled_start_date = datetime(year=2022, month=1, day=1, hour=18, minute=00),
+      scheduled_end_date = datetime(year=2022, month=1, day=1, hour=20, minute=00),
+    )
+
+    maintenance.scope = Scope(name="test")
+    maintenance.service = Service(name="test")
+
+
+    calendar = CalendarEvent.from_maintenance(maintenance)
+
+    ics_data = calendar.render()
+
+    assert "BEGIN:VCALENDAR" in str(ics_data)
+    assert "SUMMARY:TEST maintenance" in str(ics_data)